+++ examples/mux-sandbox/src/app/(payload)/admin/importMap.js
@@ -1,0 +1,8 @@
+import { MuxUploaderField as MuxUploaderField_a } from '@pro-laico/payload-mux/components/MuxUploaderField'
+import { MuxVideoGifCell as MuxVideoGifCell_b } from '@pro-laico/payload-mux/components/MuxVideoGifCell'
+
+/** @type import('payload').ImportMap */
+export const importMap = {
+  '@pro-laico/payload-mux/components/MuxUploaderField#MuxUploaderField': MuxUploaderField_a,
+  '@pro-laico/payload-mux/components/MuxVideoGifCell#MuxVideoGifCell': MuxVideoGifCell_b,
+}